Split bedticket controller logic to separate files
